# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: Listen API
  version: 0.1.0
  description: |
    Listen syncs meeting transcripts to your TinyCloud space.

    **Two-layer auth:**
    1. Bearer JWT from OpenKey (OAuth 2.1 PKCE + passkeys)
    2. TinyCloud PortableDelegation granting scoped backend access
servers:
  - url: http://localhost:3001
    description: Local development

# ── Security ────────────────────────────────────────────────────────
security:
  - bearerAuth: []

# ── Paths ───────────────────────────────────────────────────────────
paths:
...
  /api/delegations:
    post:
      operationId: sendDelegation
      summary: Submit a delegation to the backend
      description: |
        Receive a serialized PortableDelegation, verify it, and store it.
        The delegation is activated when the node is available; activation may
        remain pending after the durable row is accepted.
      tags: [Delegations]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serialized]
              properties:
                serialized:
                  type: string
                  description: Serialized PortableDelegation string
      responses:
        "200":
          description: Delegation stored; activation may be active or pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationResponse"
        "400":
          description: Invalid body or delegation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                invalidBody:
                  value:
                    error: invalid_body
                    message: "Request body must include a 'serialized' string field"
                invalidDelegation:
                  value:
                    error: invalid_delegation
                    message: Failed to process delegation
        "503":
          description: Delegation storage is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: A newer delegation or removal superseded this activation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfRejected"
        "429":
          $ref: "#/components/responses/DelegationRateLimited"

    delete:
      operationId: revokeDelegation
      summary: Remove the stored delegation copy
      description: Removes the backend's stored delegation copy and evicts its cache entry. It does not revoke an already-issued cryptographic delegation.
      tags: [Delegations]
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: Stored delegation copy removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationResponse"
              example:
                status: none
                expiresAt: null
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfRejected"
        "429":
          $ref: "#/components/responses/DelegationRateLimited"
        "500":
          description: Stored-copy removal failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/delegations/status:
    get:
      operationId: getDelegationStatus
      summary: Check delegation status
      description: |
        Returns the durable delegation status for the authenticated user without
        deleting expired or policy-stale rows.
      tags: [Delegations]
      responses:
        "200":
          description: Delegation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelegationResponse"
              examples:
                active:
                  value:
                    status: active
                    expiresAt: "2026-03-28T12:00:00.000Z"
                expired:
                  value:
                    status: expired
                    expiresAt: "2026-03-20T12:00:00.000Z"
                none:
                  value:
                    status: none
                    expiresAt: null
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/DelegationRateLimited"
        "500":
          description: Delegation status read failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OpenKey JWT access token (OAuth 2.1 PKCE)
  parameters:
    CsrfHeader:
      name: X-Requested-With
      in: header
      required: true
      description: CSRF protection header (required for state-changing methods)
      schema:
        type: string
        default: Listen
  schemas:
    DelegationResponse:
      type: object
      required: [status, expiresAt]
      properties:
        status:
          type: string
          enum: [active, expired, stale, none]
        expiresAt:
          type: [string, "null"]
          format: date-time
        activation:
          type: string
          enum: [active, pending, failed, unknown]
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error description
...
  responses:
    Unauthenticated:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: missing_token
            message: Authorization header is required
    CsrfRejected:
      description: Missing or invalid X-Requested-With header
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: csrf_rejected
            message: Missing or invalid X-Requested-With header
...
    DelegationRateLimited:
      description: Too many delegation requests (10 per 15 minutes)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: rate_limited
            message: Too many delegation requests
...